\name{rrdf-package}
\alias{rrdf-package}
\alias{rrdf}
\docType{package}
\title{
Functionality to work with RDF data.
}
\description{
Methods to read and write RDF, and query RDF using SPARQL.
}
\details{
\tabular{ll}{
Package: \tab rrdf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-03-22\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}
}
\author{
Egon Willighagen

Maintainer: Egon Willighagen <egon.willighagen@gmail.com>
}
\keyword{ package }
\keyword{ RDF }
\keyword{ SPARQL }
