\name{summarize.rdf}
\alias{summarize.rdf}
\title{
Summarized the content of a Java Model.
}
\description{
Summarized the content of a Java Model, including the number
of triples.
}
\usage{
summarize.rdf(model)
}
\arguments{
  \item{model}{
A Java Model object.
}
}
\value{
Results are returned to the concole.
}
\author{
Egon Willighagen
}
\keyword{ RDF }
\keyword{ triple }
