\name{tripleCount}
\alias{tripleCount}
\title{
Summarized the content of a Java Model.
}
\description{
Summarized the content of a Java Model, including the number
of triples.
}
\usage{
tripleCount(model)
}
\arguments{
  \item{model}{
A Java Model object.
}
}
\value{
Results are returned to the console.
}
\author{
Egon Willighagen
}
\examples{
store = new.rdf()
tripleCount(store)
}
\keyword{ RDF }
\keyword{ triple }
