\name{new.rdf}
\alias{new.rdf}
\title{
Create a new RDF triple store object
}
\description{
Create a new RDF triple store object.
}
\usage{
new.rdf()
}
\value{
A Java Model object containing the triples loaded from the file.
}
\author{
Egon Willighagen
}
\examples{
store = new.rdf()
}
\keyword{ RDF }
\keyword{ triple }
