\name{rrdf-package}
\alias{rrdf-package}
\alias{rrdf}
\docType{package}
\title{
Functionality to work with RDF data.
}
\description{
Methods to read and write RDF, and query RDF using SPARQL.
}
\details{
Package based on the Jena RDF libraries to bring RDF and SPARQL
functionality into the R environment.
}
\author{
Egon Willighagen

Maintainer: Egon Willighagen <egon.willighagen@gmail.com>
}
\keyword{ package }
\keyword{ RDF }
\keyword{ SPARQL }
